Some security improvements

+++ FinanceAPI/Controllers/UsersController.cs
@@ -1,8 +1,6 @@
 using FinanceAPI.DTOs.ApiKeys;
 using FinanceAPI.DTOs.Users;
-using FinanceAPI.Interfaces.Repositories;
 using FinanceAPI.Interfaces.Services;
-using FinanceAPI.Models;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -14,12 +12,10 @@ namespace FinanceAPI.Controllers;
 public class UsersController : AuthenticatedControllerBase
 {
     private readonly IUserService _userService;
-    private readonly IUserRepository _userRepo;
 
-    public UsersController(IUserService userService, IUserRepository userRepo)
+    public UsersController(IUserService userService)
     {
         _userService = userService;
-        _userRepo = userRepo;
     }
 
     [HttpGet]
@@ -43,6 +39,11 @@ public class UsersController : AuthenticatedControllerBase
     [HttpDelete("{userId:int}")]
     public async Task<IActionResult> Delete(int userId)
     {
+        if (userId == UserId)
+        {
+            throw new InvalidOperationException("You cannot delete your own account.");
+        }
+
         await _userService.DeleteAsync(userId);
         return NoContent();
     }
@@ -64,12 +65,7 @@ public class UsersController : AuthenticatedControllerBase
     [HttpPut("{userId:int}/password")]
     public async Task<IActionResult> SetPassword(int userId, [FromBody] AdminSetPasswordRequest request)
     {
-        User user = await _userRepo.GetByIdAsync(userId)
-                   ?? throw new KeyNotFoundException($"User {userId} not found.");
-
-        string newHash = BCrypt.Net.BCrypt.HashPassword(request.NewPassword, workFactor: 12);
-        await _userRepo.UpdatePasswordAsync(user.Id, newHash);
-
+        await _userService.AdminSetPasswordAsync(userId, request.NewPassword);
         return NoContent();
     }
